make catalog submodule
