+++ optd-persistent/src/migrator/catalog/m20241029_000001_database_metadata.rs
@@ -1,5 +1,4 @@
-use sea_orm_migration::prelude::*;
-use sea_orm_migration::schema::*;
+use sea_orm_migration::{prelude::*, schema::*};
 
 #[derive(Iden)]
 pub enum DatabaseMetadata {
