+++ optd-persistent/src/migrator/catalog/m20241029_000001_index.rs
@@ -13,9 +13,8 @@ Table index {
 }
  */
 
-use crate::migrator::cost_model::table_metadata::TableMetadata;
-use sea_orm_migration::prelude::*;
-use sea_orm_migration::schema::*;
+use crate::migrator::catalog::table_metadata::TableMetadata;
+use sea_orm_migration::{prelude::*, schema::*};
 
 #[derive(Iden)]
 pub enum Index {
