+++ optd-persistent/src/migrator/catalog/m20241029_000001_namespace_metadata.rs
@@ -1,6 +1,5 @@
-use crate::migrator::cost_model::database_metadata::DatabaseMetadata;
-use sea_orm_migration::prelude::*;
-use sea_orm_migration::schema::*;
+use crate::migrator::catalog::database_metadata::DatabaseMetadata;
+use sea_orm_migration::{prelude::*, schema::*};
 
 #[derive(Iden)]
 pub enum NamespaceMetadata {
