+++ optd-persistent/src/migrator/catalog/mod.rs
@@ -1,0 +1,19 @@
+pub(crate) mod m20241029_000001_constraint;
+pub(crate) mod m20241029_000001_constraint_attribute_junction;
+pub(crate) mod m20241029_000001_database_metadata;
+pub(crate) mod m20241029_000001_foreign_constraint_ref_attribute_junction;
+pub(crate) mod m20241029_000001_index;
+pub(crate) mod m20241029_000001_namespace_metadata;
+pub(crate) mod m20241029_000001_table_attribute;
+pub(crate) mod m20241029_000001_table_metadata;
+pub(crate) mod m20241029_000001_trigger;
+
+pub(crate) use m20241029_000001_constraint as constraint;
+pub(crate) use m20241029_000001_constraint_attribute_junction as constraint_attribute_junction;
+pub(crate) use m20241029_000001_database_metadata as database_metadata;
+pub(crate) use m20241029_000001_foreign_constraint_ref_attribute_junction as foreign_constraint_ref_attribute_junction;
+pub(crate) use m20241029_000001_index as index;
+pub(crate) use m20241029_000001_namespace_metadata as namespace_metadata;
+pub(crate) use m20241029_000001_table_attribute as table_attribute;
+pub(crate) use m20241029_000001_table_metadata as table_metadata;
+pub(crate) use m20241029_000001_trigger as trigger;
