+++ optd-persistent/src/migrator/cost_model/m20241029_000001_attribute_stat.rs
@@ -12,8 +12,7 @@ Table attribute_stat {
 } */
 
 use crate::migrator::cost_model::event::Event;
-use sea_orm_migration::prelude::*;
-use sea_orm_migration::schema::*;
+use sea_orm_migration::{prelude::*, schema::*};
 
 #[derive(Iden)]
 pub enum AttributeStat {
