+++ optd-persistent/src/migrator/memo/mod.rs
@@ -1,21 +1,17 @@
 pub(crate) mod m20241029_000001_cascades_group;
 pub(crate) mod m20241029_000001_group_winner;
-
 pub(crate) mod m20241029_000001_logical_expression;
 pub(crate) mod m20241029_000001_logical_group_junction;
 pub(crate) mod m20241029_000001_logical_property;
-
 pub(crate) mod m20241029_000001_physical_expression;
 pub(crate) mod m20241029_000001_physical_group_junction;
 pub(crate) mod m20241029_000001_physical_property;
 
 pub(crate) use m20241029_000001_cascades_group as cascades_group;
 pub(crate) use m20241029_000001_group_winner as group_winner;
-
 pub(crate) use m20241029_000001_logical_expression as logical_expression;
 pub(crate) use m20241029_000001_logical_group_junction as logical_group_junction;
 pub(crate) use m20241029_000001_logical_property as logical_property;
-
 pub(crate) use m20241029_000001_physical_expression as physical_expression;
 pub(crate) use m20241029_000001_physical_group_junction as physical_group_junction;
 pub(crate) use m20241029_000001_physical_property as physical_property;
